requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
pinecone-client==5.0.1
anthropic==0.42.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
tiktoken==0.8.0
chardet==5.2.0
pytest==8.3.4
